% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPredictions.R
\name{addPredictions}
\alias{addPredictions}
\title{Add model predictions to the original data set}
\usage{
addPredictions(results, data)
}
\arguments{
\item{results}{An object of class trainCV that contains the following
elements: models, predictions, and testSets.}

\item{data}{A data frame.}
}
\value{
Returns a data frame with predictions added as the last column.
}
\description{
A function that adds the predictions obtained from the trainCV function to
data.
}
\examples{
# Using rbps data set available with package
imputedRBPs <- impute(rbps, "mean")

# Changing the class of hasCanonicalRBDs (an independent variable) to
# numeric to be used for the trainCV function
imputedRBPs$hasCanonicalRBDs <- as.numeric(imputedRBPs$hasCanonicalRBDs)

# Remove the Human Gene and pLI columns
imputedRBPs <- subset(imputedRBPs, select = -c(1, 8))

results <- trainCV(data = imputedRBPs, colIndex = 10)
newRBPs <- addPredictions(results, rbps)
head(newRBPs)

}
