% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{impute}
\alias{impute}
\title{Imputation of missing data}
\usage{
impute(data, replace)
}
\arguments{
\item{data}{A data frame to be imputated.}

\item{replace}{A value used to replace the missing values in a column.
Should be one of the mean, median, or mode.}
}
\value{
Returns the modified data frame data.
}
\description{
A function that replaces the missing values in a column with values imputed
from the observed data (i.e. the mean, median, or mode for that variable).
}
\examples{
# Using rbps data set available with package
# Replace the missing values in rbps with the mean
imputated_rbps <- impute(rbps, "mean")
head(imputated_rbps)

}
\references{
Little, JA. R. & Rubin B. D. (1987). \emph{Statistical analysis of missing
data}. John Wiley & Sons.
}
