% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingValues.R
\name{MissingValues}
\alias{MissingValues}
\title{Checking for missing values in the data set}
\usage{
MissingValues(data)
}
\arguments{
\item{data}{A data frame to be analyzed.}
}
\value{
Returns an object of class MissingValues with results.
\itemize{
  \item NA_results - The number of NA values in each column.
  \item unique_results - The number of unique values in each column.
}
}
\description{
A function that checks the number of NA values in each column and produces
a visual output for the missing values in the data set. The most missing
variable is farthest to the left in the visual output.
}
\examples{
# Using rbps data set available with package

# Check for missing values in the rbps data set
MissingValuesResults <- MissingValues(data = rbps)
MissingValuesResults$NA_results

}
\references{

}
