% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plots the Receiver Operating Characteristics Curve (ROC)}
\usage{
plotROC(pred, truth, ...)
}
\arguments{
\item{pred}{A vector, list, matrix, or data frame containing the predictions
from logistic regression.}

\item{truth}{A vector, list, matrix, or data frame containing the true class
labels. The dimensions must be the same as predictions.}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\value{
Returns a plot of an ROC curve with its AUC value.
}
\description{
A function that computes the area under the ROC curve (AUC) from model
predictions and plots the ROC.
}
\examples{
# Using rbps data set available with package
data(rbps)
imputedRBPs <- impute(rbps, "mean")
imputedRBPs$hasCanonicalRBDs <- as.numeric(imputedRBPs$hasCanonicalRBDs)

# Remove the Human Gene and pLI columns
rbps <- subset(imputedRBPs, select = -c(1, 8))

results <- trainCV(data = rbps, colIndex = 10)
plotROC(results$predictions[[1]], results$testSets[[1]]$autism_genes)

}
\references{
Zach. (2021, September 9). \emph{How to calculate AUC (area under curve) in
R.} Statology. \href{https://www.statology.org/auc-in-r/}{Link}.

Bradley, A. P. (1997). The use of the area under the ROC curve in the
evaluation of machine learning algorithms. \emph{Pattern Recognition}, 30(7),
1145–1159. \href{https://doi.org/10.1016/s0031-3203(96)00142-2}{Link}.
}
