% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessData.R
\name{preProcessData}
\alias{preProcessData}
\title{Pre-process the data set to be used for logistic regression modelling}
\usage{
preProcessData(data, dependentVarIndex, deleteColumns = NULL)
}
\arguments{
\item{data}{A data frame to be pre-processed.}

\item{dependentVarIndex}{A positive integer indicating the index of the
dependent variable column in the data set.}

\item{deleteColumns}{A numeric vector containing the columns to be deleted.
These are user-specified indices of the columns that will not be used for
building the logistic regression model with the tranCV function. An
optional parameter.}
}
\value{
Returns the pre-processed data frame.
}
\description{
A function that pre-processes the data for use in building the logistic
regression model.
}
\examples{
# Using rbps data set available with package
data(rbps)
imputedRBPs <- impute(data = rbps, replace = "mean")

newDat <- preProcessData(data = imputedRBPs,
                         dependentVarIndex = 12,
                         deleteColumns = c(1, 8))

head(newDat)

}
