% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneCTLR.R
\name{runGeneCTLR}
\alias{runGeneCTLR}
\title{Launch Shiny App for GeneCTLR}
\usage{
runGeneCTLR()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for GeneCTLR.
}
\examples{
\dontrun{

GeneCTLR::runGeneCTLR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
