% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPredictions.R
\name{addPredictions}
\alias{addPredictions}
\title{Add model predictions to the original data set}
\usage{
addPredictions(results, data)
}
\arguments{
\item{results}{An object of class trainCV that contains the following
elements: models, predictions, and test_sets.}

\item{data}{A data frame.}
}
\value{
Returns a data frame with predictions added as the last column.
}
\description{
A function that adds the predictions obtained from the trainCV function to
data.
}
\examples{
# Using rbps data set available with package
imputated_rbps <- impute(rbps, "mean")

# Changing the class of hasCanonicalRBDs (an independent variable) to
# numeric to be used for the trainCV function
imputated_rbps$hasCanonicalRBDs <- as.numeric(imputated_rbps$hasCanonicalRBDs)

# Remove the Human Gene and pLI columns
imputated_rbps <- subset(imputated_rbps, select = -c(1, 8))

rbps_results <- trainCV(data = imputated_rbps, col_index = 10)
new_rbps <- AddPredictions(rbps_results, rbps)
head(new_rbps)

}
