% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Imputation of missing data}
\usage{
impute(data, replace)
}
\arguments{
\item{data}{A data frame to be imputed.}

\item{replace}{A value used to replace the missing values in a column.
Should be either the mean or the median.}
}
\value{
Returns the modified data frame data.
}
\description{
A function that replaces the missing values in a column with values imputed
from the observed data (i.e. the mean, median, or mode for that variable).
}
\details{
Only variables of class numeric that contain NA values will be
imputed.
}
\examples{
# Using rbps data set available with package
data(rbps)
# Replace the missing values in rbps with the mean
imputedRBPs <- impute(rbps, "mean")
head(imputedRBPs)

}
\references{
Little, JA. R. & Rubin B. D. (1987). \emph{Statistical analysis of missing
data}. John Wiley & Sons.
}
