% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPR.R
\name{plotPR}
\alias{plotPR}
\title{Plots the Precision-Recall (PR) Curve}
\usage{
plotPR(results, dependentVarIndex, title)
}
\arguments{
\item{results}{an S3 object of class trainCV with the following elements:
\itemize{
\item models - Fitted logistic regression models.
\item predictions - The probabilities predicted by each model.
\item testSets - Test data sets used in each fold of the cross validation.}}

\item{dependentVarIndex}{A positive integer indicating the index of the
dependent variable column in the data set.}

\item{title}{A string of characters denoting the title of the plot. An
optional parameter. The default value, "Precision-Recall Curve", is used if
there isn't a user-specified value for title.}
}
\value{
Returns a precision-recall curve with AUCPR values demonstrating
the performance of the algorithm
}
\description{
A function that computes the area under the precision-recall (PR) curve from
model predictions and plots the precision-recall curve. PR curves are
particularly useful for imbalanced data sets where we observe one class more
frequently than the other class.
}
\examples{
# Using rbps data set available with package
data(rbps)
imputedRBPs <- impute(data = rbps, replace = "mean")

newDat <- preProcessData(data = imputedRBPs,
                         dependentVarIndex = 12,
                         deleteColumns = c(1, 8))

# Using default K = 5
results <- trainCV(data = newDat, dependentVarIndex = 10)

plotPR(results = results,
       dependentVarIndex = 10,
       title = "Precision-Recall Curves for Autism Gene Prediction")

}
\references{
Boyd, K., Eng, K. H., & Page, C. D. (2013). Area under the
precision-recall curve: Point estimates and confidence intervals.
\emph{Advanced Information Systems Engineering}, 451–466.
\href{https://doi.org/10.1007/978-3-642-40994-3_29}{Link}

Fu, G. H., Yi, L. Z., & Pan, J. (2018). Tuning model parameters in
class‐imbalanced learning with precision‐recall curve. \emph{Biometrical
Journal}, 61(3), 652–664. \href{https://doi.org/10.1002/bimj.201800148}{Link}

Williams, C. K. (2021). The effect of class imbalance on precision-recall
curves. \emph{Neural Computation}, 33(4), 853–857.
\href{https://doi.org/10.1162/neco_a_01362}{Link}.
}
