% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPR.R
\name{plotPR}
\alias{plotPR}
\title{Plots the Precision-Recall Curve (PRC)}
\usage{
plotPR(pred, truth, ...)
}
\arguments{
\item{pred}{A vector, list, matrix, or data frame containing the predictions
from logistic regression.}

\item{truth}{A vector, list, matrix, or data frame containing the true class
labels. The dimensions must be the same as predictions.}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\value{
Returns a plot of a precision-recall curve with its AUC value.
}
\description{
A function that computes the area under the precision-recall curve (PRC) from
model predictions and plots the precision-recall curve.
}
\examples{
# Using rbps data set available with package
data(rbps)
imputedRBPs <- impute(rbps, "mean")
imputedRBPs$hasCanonicalRBDs <- as.numeric(imputedRBPs$hasCanonicalRBDs)

# Remove the Human Gene and pLI columns
rbps <- subset(imputedRBPs, select = -c(1, 8))

results <- trainCV(data = rbps, colIndex = 10)
plotPR(results$predictions[[3]], results$testSets[[3]]$autism_genes)

}
\references{
Fu, G. H., Yi, L. Z., & Pan, J. (2018). Tuning model parameters in
class‐imbalanced learning with precision‐recall curve. \emph{Biometrical
Journal}, 61(3), 652–664. \href{https://doi.org/10.1002/bimj.201800148}{Link}

Williams, C. K. (2021). The effect of class imbalance on precision-recall
curves. \emph{Neural Computation}, 33(4), 853–857.
\href{https://doi.org/10.1162/neco_a_01362}{Link}.
}
