% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plots the Receiver Operating Characteristics (ROC) Curve}
\usage{
plotROC(results, dependentVarIndex)
}
\arguments{
\item{results}{an S3 object of class trainCV with the following elements:
\itemize{
\item models - Fitted logistic regression models.
\item predictions - The probabilities predicted by each model.
\item testSets - Test data sets used in each fold of the cross validation.}}

\item{dependentVarIndex}{A positive integer indicating the index of the
dependent variable column in the data set.}
}
\value{
Returns an ROC curve with AUC values showing the possibility that
the model ranks a random positive event more highly than a random negative
event.
}
\description{
A function that computes the area under the ROC curve (AUC) from model
predictions and plots the ROC curve.
}
\examples{
# Using rbps data set available with package
data(rbps)
imputedRBPs <- impute(data = rbps, replace = "mean")

newDat <- preProcessData(data = imputedRBPs,
                         dependentVarIndex = 12,
                         deleteColumns = c(1, 8))

# Using default K = 5
results <- trainCV(data = newDat, dependentVarIndex = 10)

plotROC(results = results,
        dependentVarIndex = 10)

}
\references{
Bradley, A. P. (1997). The use of the area under the ROC curve in the
evaluation of machine learning algorithms. \emph{Pattern Recognition}, 30(7),
1145–1159. \href{https://doi.org/10.1016/s0031-3203(96)00142-2}{Link}.

Google. (n.d.). \emph{Classification: Roc curve and AUC; machine learning crash
course; google developers.} Google.
\href{https://developers.google.com/machine-learning/crash-course/classification/roc-and-auc.}{Link}.

Zach. (2021, September 9). \emph{How to calculate AUC (area under curve) in
R.} Statology. \href{https://www.statology.org/auc-in-r/}{Link}.
}
