% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainCV.R
\name{trainCV}
\alias{trainCV}
\title{Performs logistic model training and cross validation}
\usage{
trainCV(data, colIndex, K = 5)
}
\arguments{
\item{data}{A data frame to be trained. The data frame should only contain
one dependent variable and one or more independent variables. All independent
variables to be used in training the model should be of type numeric.}

\item{colIndex}{A positive integer indicating the index of the dependent
variable column in the data set.}

\item{K}{A positive integer indicating the number of groups that a given
data set is to be split into.}
}
\value{
Returns an S3 object of class trainCV with the following elements:
\itemize{
\item models - Fitted logistic regression models.
\item predictions - The probabilities predicted by each model.
\item testSets - Test data sets used in each fold of the cross validation.
}
}
\description{
A function that trains the logistic regression model and performs a K-fold
cross validation.
}
\examples{
# Changing the class of hasCanonicalRBDs (an independent variable) to
# numeric to be used for the trainCV function
imputedRBPs <- impute(rbps, "mean")
imputedRBPs$hasCanonicalRBDs <- as.numeric(imputedRBPs$hasCanonicalRBDs)

# Remove the Human Gene and pLI columns
rbps <- subset(imputedRBPs, select = -c(1, 8))

results <- trainCV(data = rbps, colIndex = 10)

}
\references{
Alice, M. (2020, July 5). \emph{How to perform a logistic regression in R:
R-bloggers.} R. \href{https://www.r-bloggers.com/2015/09/how-to-perform-a-logistic-regression-in-r/}{Link}.

Selva, P. (n.d.). \emph{Logistic Regression with R.} r-statistics.co.
\href{http://r-statistics.co/Logistic-Regression-With-R.html}{Link}.
}
